requirements.txt
fastapi
uvicorn
groq
python-multipart
pydantic
gtts
python-dotenv